+++ src/main/java/malek/mod_science/power/IPowerCarrier.java
@@ -3,8 +3,8 @@ package malek.mod_science.power;
 import net.minecraft.util.math.BlockPos;
 
 public interface IPowerCarrier {
-    double getFireEfficiency();
-    double getLightEfficiency();
-    double getArcEfficiency();
-    double getTimeEfficiency();
+    int getFireEfficiency();
+    int getLightEfficiency();
+    int getArcEfficiency();
+    int getTimeEfficiency();
 }
